+++ NEWS.md
@@ -1,4 +1,5 @@
 # Version 2.2.2.9000
+* Added `strip0()` (#41).
 * Added `snakeToCamel()`, `medianNA()`, `rangeNA()` (#32).
 * Reduced test_regTable.R reliance on multiwayvcov package (#36).
 * Minor update to `qw()` documentation (#37).